requirements.txt
fastapi
uvicorn
pyserial
python-multipart
pytest

